+++ R/rRofex.R
@@ -4,5 +4,7 @@
 #' @importFrom tibble enframe
 #' @importFrom tidyr separate
 #' @importFrom magrittr %>%
-#' @importFrom glue glue
+#' @importFrom glue glue glue_collapse
+#' @importFrom purrr some
+#' @importFrom rlang is_null
 "_PACKAGE"